Move timing sensitive tests into repeatable mode
