+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/queries/AsNodeOperatorQueriesTest.java
@@ -3,39 +3,27 @@ package com.hedera.services.bdd.suites.queries;
 
 import static com.hedera.services.bdd.junit.ContextRequirement.THROTTLE_OVERRIDES;
 import static com.hedera.services.bdd.junit.TestTags.CRYPTO;
-import static com.hedera.services.bdd.spec.HapiSpec.customizedHapiTest;
 import static com.hedera.services.bdd.spec.HapiSpec.hapiTest;
-import static com.hedera.services.bdd.spec.assertions.AccountInfoAsserts.changeFromSnapshot;
-import static com.hedera.services.bdd.spec.assertions.AccountInfoAsserts.lessThan;
-import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountBalance;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountInfo;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getContractBytecode;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getContractInfo;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getFileContents;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getFileInfo;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getScheduleInfo;
-import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTokenInfo;
-import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTopicInfo;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCreate;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.createTopic;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoTransfer;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.fileCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.scheduleCreate;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.uploadInitCode;
 import static com.hedera.services.bdd.spec.transactions.crypto.HapiCryptoTransfer.tinyBarsFromTo;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.balanceSnapshot;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.inParallel;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.newKeyNamed;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.overridingThrottles;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.sleepFor;
 import static com.hedera.services.bdd.suites.HapiSuite.DEFAULT_PAYER;
-import static com.hedera.services.bdd.suites.HapiSuite.ONE_HUNDRED_HBARS;
 import static com.hedera.services.bdd.suites.HapiSuite.flattened;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.BUSY;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.OK;
-import static com.hederahashgraph.api.proto.java.TokenType.NON_FUNGIBLE_UNIQUE;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.hedera.services.bdd.junit.HapiTest;
@@ -45,14 +33,12 @@ import com.hedera.services.bdd.junit.OrderedInIsolation;
 import com.hedera.services.bdd.junit.hedera.HederaNode;
 import com.hedera.services.bdd.junit.support.TestLifecycle;
 import com.hedera.services.bdd.spec.HapiSpecOperation;
-import com.hedera.services.bdd.suites.regression.system.LifecycleTest;
 import com.hederahashgraph.api.proto.java.AccountAmount;
 import com.hederahashgraph.api.proto.java.AccountID;
 import com.hederahashgraph.api.proto.java.CryptoGetInfoQuery;
 import com.hederahashgraph.api.proto.java.CryptoTransferTransactionBody;
 import com.hederahashgraph.api.proto.java.Query;
 import com.hederahashgraph.api.proto.java.ResponseCodeEnum;
-import com.hederahashgraph.api.proto.java.TokenSupplyType;
 import com.hederahashgraph.api.proto.java.Transaction;
 import com.hederahashgraph.api.proto.java.TransactionBody;
 import com.hederahashgraph.api.proto.java.TransferList;
@@ -65,7 +51,6 @@ import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import java.util.stream.IntStream;
@@ -74,6 +59,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.condition.EnabledIfEnvironmentVariable;
 
 /**
  * A class with Node Operator Queries tests
@@ -82,7 +68,7 @@ import org.junit.jupiter.api.Tag;
 @DisplayName("Node Operator Queries")
 @HapiTestLifecycle
 @OrderedInIsolation
-public class AsNodeOperatorQueriesTest extends NodeOperatorQueriesBase implements LifecycleTest {
+public class AsNodeOperatorQueriesTest extends NodeOperatorQueriesBase {
 
     private static final int BURST_SIZE = 20;
 
@@ -98,277 +84,6 @@ public class AsNodeOperatorQueriesTest extends NodeOperatorQueriesBase implement
     static void beforeAll(@NonNull final TestLifecycle lifecycle) {
         lifecycle.doAdhoc(createAllAccountsAndTokens());
         nodes = lifecycle.getNodes();
-    }
-
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorQueryVerifyPayerBalanceForAccountBalance() {
-        return customizedHapiTest(
-                Map.of("memo.useSpecName", "false"),
-                cryptoCreate(NODE_OPERATOR).balance(ONE_HUNDRED_HBARS),
-                cryptoCreate(PAYER).balance(ONE_HUNDRED_HBARS),
-                // perform getAccountBalance() query, pay for the query with payer account
-                getAccountBalance(NODE_OPERATOR).payingWith(PAYER),
-                sleepFor(3_000),
-                // assert payer is charged
-                getAccountBalance(PAYER).hasTinyBars(ONE_HUNDRED_HBARS),
-                // perform free query to local port with asNodeOperator() method
-                getAccountBalance(NODE_OPERATOR).payingWith(PAYER).asNodeOperator(),
-                sleepFor(3_000),
-                // assert payer is not charged as the query is performed as node operator
-                getAccountBalance(PAYER).hasTinyBars(ONE_HUNDRED_HBARS));
-    }
-
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorQueryVerifyPayerBalanceForAccountInfo() {
-        return customizedHapiTest(
-                Map.of("memo.useSpecName", "false"),
-                flattened(
-                        nodeOperatorAccount(),
-                        payerAccount(),
-                        balanceSnapshot("payerInitialBalance", PAYER),
-                        // perform getAccountInfo() query, pay for the query with payer account
-                        // the grpc client performs the query to different ports
-                        getAccountInfo(NODE_OPERATOR).payingWith(PAYER),
-                        sleepFor(3_000),
-                        // assert payer is charged
-                        getAccountBalance(PAYER).hasTinyBars(changeFromSnapshot("payerInitialBalance", -QUERY_COST)),
-                        // perform free query to local port with asNodeOperator() method
-                        getAccountInfo(NODE_OPERATOR).payingWith(PAYER).asNodeOperator(),
-                        sleepFor(3_000),
-                        // assert payer is not charged as the query is performed as node operator
-                        getAccountBalance(PAYER).hasTinyBars(changeFromSnapshot("payerInitialBalance", -QUERY_COST))));
-    }
-
-    /**
-     * AccountInfoQuery
-     *  1.) Tests that verifies the payer balance is not charged when a node operator AccountInfoQuery is performed.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorAccountInfoQueryNotCharged() {
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                getAccountInfo(NODE_OPERATOR).payingWith(NODE_OPERATOR).asNodeOperator(),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS)));
-    }
-
-    /**
-     * AccountInfoQuery
-     *  2.)  Tests that a signed transaction is not required for the query if it is performed as node operator.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorAccountInfoQueryNotSigned() {
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                getAccountInfo(NODE_OPERATOR)
-                        .payingWith(NODE_OPERATOR)
-                        .signedBy(DEFAULT_PAYER)
-                        .asNodeOperator(),
-                sleepFor(3000),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS)));
-    }
-
-    /**
-     * AccountInfoQuery
-     *  3.) Tests that verifies the payer balance is charged when a AccountInfoQuery is performed.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorAccountInfoQueryCharged() {
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                getAccountInfo(NODE_OPERATOR).payingWith(NODE_OPERATOR).via("accountInfoQueryTxn"),
-                sleepFor(2000),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(lessThan(ONE_HUNDRED_HBARS))));
-    }
-
-    /**
-     * AccountBalanceQuery
-     *  1.) Tests that verifies the payer balance is not charged(free query) when a node operator AccountBalanceQuery is performed.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorAccountBalanceQueryNotCharged() {
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                getAccountBalance(NODE_OPERATOR).payingWith(NODE_OPERATOR).asNodeOperator(),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS)));
-    }
-
-    /**
-     * AccountBalanceQuery
-     *  2.)  Tests that a signed transaction is not required for the query if it is performed as node operator.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorAccountBalanceQueryNotSigned() {
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                getAccountBalance(NODE_OPERATOR)
-                        .payingWith(NODE_OPERATOR)
-                        .signedBy(DEFAULT_PAYER)
-                        .asNodeOperator(),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS)));
-    }
-
-    /**
-     * TokenInfoQuery - FT
-     *  1.) Tests that verifies the payer balance is not charged when a node operator TokenInfoQuery is performed.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorTokenInfoQueryNotCharged() {
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                tokenCreate(FUNGIBLE_QUERY_TOKEN),
-                getTokenInfo(FUNGIBLE_QUERY_TOKEN).payingWith(NODE_OPERATOR).asNodeOperator(),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS)));
-    }
-
-    /**
-     * TokenInfoQuery - FT
-     *  2.) Tests that verifies the payer balance is charged when a AccountInfoQuery is performed.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorTokenInfoQueryCharged() {
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                tokenCreate(ANOTHER_FUNGIBLE_QUERY_TOKEN),
-                getTokenInfo(ANOTHER_FUNGIBLE_QUERY_TOKEN).payingWith(NODE_OPERATOR),
-                sleepFor(3000),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(lessThan(ONE_HUNDRED_HBARS))));
-    }
-
-    /**
-     * TokenInfoQuery - FT
-     *  3.) Tests that a signed transaction is not required for the query if it is performed as node operator.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorTokenInfoQuerySignature() {
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                tokenCreate(ANOTHER_FUNGIBLE_QUERY_TOKEN),
-                getTokenInfo(ANOTHER_FUNGIBLE_QUERY_TOKEN)
-                        .payingWith(NODE_OPERATOR)
-                        .asNodeOperator(),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS)));
-    }
-
-    /**
-     * TokenInfoQuery - NFT
-     *  1.) Tests that verifies the payer balance is not charged when a node operator TokenInfoQuery is performed.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorTokenInfoQueryNftNotCharged() {
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                newKeyNamed(SUPPLY_KEY),
-                newKeyNamed(WIPE_KEY),
-                cryptoCreate(TREASURY),
-                tokenCreate(NON_FUNGIBLE_TOKEN)
-                        .supplyType(TokenSupplyType.FINITE)
-                        .tokenType(NON_FUNGIBLE_UNIQUE)
-                        .treasury(TREASURY)
-                        .maxSupply(12L)
-                        .wipeKey(WIPE_KEY)
-                        .supplyKey(SUPPLY_KEY)
-                        .initialSupply(0L),
-                getTokenInfo(NON_FUNGIBLE_TOKEN).payingWith(NODE_OPERATOR).asNodeOperator(),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS)));
-    }
-
-    /**
-     * TokenInfoQuery - NFT
-     *  2.) Tests that verifies the payer balance is charged when a AccountInfoQuery is performed.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorTokenInfoQueryNftCharged() {
-        return hapiTest(flattened(
-                createAllAccountsAndTokens(),
-                newKeyNamed(SUPPLY_KEY),
-                newKeyNamed(WIPE_KEY),
-                cryptoCreate(TREASURY),
-                flattened(
-                        nodeOperatorAccount(),
-                        tokenCreate(NON_FUNGIBLE_TOKEN)
-                                .supplyType(TokenSupplyType.FINITE)
-                                .tokenType(NON_FUNGIBLE_UNIQUE)
-                                .treasury(TREASURY)
-                                .maxSupply(12L)
-                                .wipeKey(WIPE_KEY)
-                                .supplyKey(SUPPLY_KEY)
-                                .initialSupply(0L),
-                        getTokenInfo(FUNGIBLE_QUERY_TOKEN).payingWith(NODE_OPERATOR),
-                        sleepFor(3000),
-                        getAccountBalance(NODE_OPERATOR).hasTinyBars(lessThan(ONE_HUNDRED_HBARS)))));
-    }
-
-    /**
-     * TokenInfoQuery - NFT
-     *  3.) Tests that a signed transaction is not required for the query if it is performed as node operator.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorTokenInfoQueryNftSignature() {
-        return hapiTest(flattened(
-                createAllAccountsAndTokens(),
-                newKeyNamed(SUPPLY_KEY),
-                newKeyNamed(WIPE_KEY),
-                cryptoCreate(TREASURY),
-                flattened(
-                        nodeOperatorAccount(),
-                        tokenCreate(NON_FUNGIBLE_TOKEN)
-                                .supplyType(TokenSupplyType.FINITE)
-                                .tokenType(NON_FUNGIBLE_UNIQUE)
-                                .treasury(TREASURY)
-                                .maxSupply(12L)
-                                .wipeKey(WIPE_KEY)
-                                .supplyKey(SUPPLY_KEY)
-                                .initialSupply(0L),
-                        getTokenInfo(FUNGIBLE_QUERY_TOKEN)
-                                .payingWith(NODE_OPERATOR)
-                                .signedBy(DEFAULT_PAYER)
-                                .asNodeOperator(),
-                        getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS))));
-    }
-
-    /**
-     * TopicInfoQuery
-     *  1.) Tests that verifies the payer balance is not charged when a node operator TopicInfoQuery is performed.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorTopicInfoQueryNotCharged() {
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                createTopic(TOPIC),
-                getTopicInfo(TOPIC).payingWith(NODE_OPERATOR).asNodeOperator(),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS)));
-    }
-
-    /**
-     * TopicInfoQuery
-     *  2.) Tests that verifies the payer balance is charged when a TopicInfoQuery is performed.
-     */
-    @HapiTest
-    final Stream<DynamicTest> nodeOperatorTopicInfoQueryCharged() {
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                createTopic(TOPIC),
-                getTopicInfo(TOPIC).payingWith(NODE_OPERATOR),
-                sleepFor(4000),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(lessThan(ONE_HUNDRED_HBARS))));
-    }
-
-    /**
-     * Get File Contents tests
-     */
-    @HapiTest
-    @DisplayName("Only node operators aren't charged for file contents queries")
-    final Stream<DynamicTest> fileGetContentsQueryNodeOperatorNotCharged() {
-        final var filename = "anyFile.txt";
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                payerAccount(),
-                fileCreate(filename).contents("anyContent"),
-                getFileContents(filename).payingWith(NODE_OPERATOR).asNodeOperator(),
-                getFileContents(filename).payingWith(PAYER),
-                sleepFor(3_000),
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS),
-                getAccountBalance(PAYER).hasTinyBars(lessThan(ONE_HUNDRED_HBARS))));
     }
 
     @HapiTest
@@ -395,27 +110,6 @@ public class AsNodeOperatorQueriesTest extends NodeOperatorQueriesBase implement
                         .hasAnswerOnlyPrecheck(ResponseCodeEnum.INVALID_SIGNATURE)));
     }
 
-    /**
-     * Get File Info tests
-     */
-    @HapiTest
-    @DisplayName("Only node operators aren't charged for file info queries")
-    final Stream<DynamicTest> getFileInfoQueryNodeOperatorNotCharged() {
-        final var filename = "anyFile.txt";
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                payerAccount(),
-                fileCreate(filename).contents("anyContentAgain").payingWith(PAYER),
-                // Both the node operator and payer submit queries
-                getFileInfo(filename).payingWith(NODE_OPERATOR).asNodeOperator(),
-                getFileInfo(filename).payingWith(PAYER),
-                sleepFor(1_000),
-                // The node operator wasn't charged
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS),
-                // But the payer was charged
-                getAccountBalance(PAYER).hasTinyBars(lessThan(ONE_HUNDRED_HBARS))));
-    }
-
     @HapiTest
     @DisplayName("Only node operators don't need to sign file info queries")
     final Stream<DynamicTest> getFileInfoQueryNoSigRequired() {
@@ -438,28 +132,6 @@ public class AsNodeOperatorQueriesTest extends NodeOperatorQueriesBase implement
                         .payingWith(PAYER)
                         .signedBy(someoneElse)
                         .hasAnswerOnlyPrecheck(ResponseCodeEnum.INVALID_SIGNATURE)));
-    }
-
-    /**
-     * Get a Smart Contract Function tests
-     */
-    @HapiTest
-    @DisplayName("Only node operators aren't charged for contract info queries")
-    final Stream<DynamicTest> getSmartContractQueryNodeOperatorNotCharged() {
-        final var contract = "PretendPair"; // any contract, nothing special about this one
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                payerAccount(),
-                uploadInitCode(contract),
-                contractCreate(contract),
-                // Both the node operator and payer submit queries
-                getContractInfo(contract).payingWith(NODE_OPERATOR).asNodeOperator(),
-                getContractInfo(contract).payingWith(PAYER),
-                sleepFor(1_000),
-                // The node operator wasn't charged
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS),
-                // But the payer was charged
-                getAccountBalance(PAYER).hasTinyBars(lessThan(ONE_HUNDRED_HBARS))));
     }
 
     @HapiTest
@@ -512,29 +184,6 @@ public class AsNodeOperatorQueriesTest extends NodeOperatorQueriesBase implement
                         .hasAnswerOnlyPrecheck(ResponseCodeEnum.INVALID_SIGNATURE)));
     }
 
-    /**
-     * Get Schedule Info tests
-     */
-    @HapiTest
-    @DisplayName("Only node operators aren't charged for schedule info queries")
-    final Stream<DynamicTest> getScheduleInfoQueryNodeOperatorNotCharged() {
-        final var txnToSchedule =
-                cryptoTransfer(tinyBarsFromTo(PAYER, DEFAULT_PAYER, 1)); // any txn, nothing special here
-        final var schedule = "anySchedule";
-        return hapiTest(flattened(
-                nodeOperatorAccount(),
-                payerAccount(),
-                scheduleCreate(schedule, txnToSchedule),
-                // Both the node operator and payer submit queries
-                getScheduleInfo(schedule).payingWith(NODE_OPERATOR).asNodeOperator(),
-                getScheduleInfo(schedule).payingWith(PAYER),
-                sleepFor(1_000),
-                // The node operator wasn't charged
-                getAccountBalance(NODE_OPERATOR).hasTinyBars(ONE_HUNDRED_HBARS),
-                // But the payer was charged
-                getAccountBalance(PAYER).hasTinyBars(lessThan(ONE_HUNDRED_HBARS))));
-    }
-
     @HapiTest
     @DisplayName("Only node operators don't need to sign schedule info queries")
     final Stream<DynamicTest> getScheduleInfoQueryNoSigRequired() {
@@ -583,6 +232,8 @@ public class AsNodeOperatorQueriesTest extends NodeOperatorQueriesBase implement
     }
 
     @HapiTest
+    // Not reliable in OS X local environment for some reason
+    @EnabledIfEnvironmentVariable(named = "CI", matches = "true")
     @DisplayName("Node Operator Submit Query not from Localhost")
     final Stream<DynamicTest> submitCryptoTransfer() {
         final Query query = buildCryptoAccountInfoQuery();
